test(ci): give the install-free lane a name and a contract
`ci-test-plan.test.mjs` was the only suite running before `npm ci`, so
every assertion that needed a bare checkout accreted there regardless of
subject: 66 tests, of which 35 read `.github/workflows/*.yml` and had
nothing to do with the test planner. The file name carried no hint that
its one hard rule is "import nothing that is not installed yet", which is
how a closure assertion needing esbuild was very nearly added to it. That
import would have thrown `ERR_MODULE_NOT_FOUND` in the sole required
context and frozen every merge in the repository.

Split by what a test reads. `ci-test-plan.test.mjs` keeps the 30 tests
that exercise the planner against an in-memory graph; the 36 that read a
workflow move to `ci-workflow-policy.test.mjs`, beside the two policy
suites that already follow that name.

Naming the constraint is not enforcing it, so derive it: the new suite
reads which files the steps above `setup-node` run and walks their
transitive local imports, failing on any specifier that is neither a
`node:` builtin nor a repository module holding the same rule. Moving a
step below the install lifts the constraint and adding one above imposes
it, with no list to maintain.

The closure assertion itself now lives in
`windows-package-source-closure.test.mjs`, which `check:release` runs
after installing, and `windows-recovery.yml` joined the release-contract
inputs so editing that filter selects the gate that checks it.

Two app-icon tests iterated a hand-written `APP_ICON_INPUTS`; the derived
test computes that same set from the suites the step runs, so they folded
into it and the list is gone. `prepare-windows-upgrade-baseline.test.mjs`
joined `check:release` without joining the planner's release set — the
guardrail that exists for exactly that caught it.

Generated-by: Claude Code
